adding postgres specific fetcher
