+++ modelmapper/example/some_model_setup.toml
@@ -16,7 +16,7 @@ string_fields_can_be_nullable = false
 training_csvs = ['../../tests/fixtures/training_fixture1.csv', '../../tests/fixtures/training_fixture1_to_xml_back_to_csv.csv']
 output_model_file = 'sample_model.py'
 ignore_lines_that_include_only_subset_of = ["", "-"]
-ignore_fields_in_signature_calculation = []
+ignore_fields_in_signature_calculation = ["id", "raw_key_id"]
 ignore_duplicate_rows_when_importing = true
 slack_username = ''
 slack_channel = ''
