+++ modelmapper/templates/setup_template.toml
@@ -16,7 +16,7 @@ string_fields_can_be_nullable = false  # Normally string fields should not be nu
 training_csvs = []  # The list of relative paths to the training csvs
 output_model_file = ""  # The relative path to the ORM model file that the output generated model will be inserted into.
 ignore_lines_that_include_only_subset_of = ["", "-"]  # Ignore lines that only include these characters
-ignore_fields_in_signature_calculation = []  # Only used when ignore_duplicate_rows_when_importing is true. Ignore these field names when calculating the signature of the row for avoiding duplicate data. Only used when importing the data into database and NOT for training the model.
+ignore_fields_in_signature_calculation = ["id", "raw_key_id"]  # Only used when ignore_duplicate_rows_when_importing is true. Ignore these field names when calculating the signature of the row for avoiding duplicate data. Only used when importing the data into database and NOT for training the model.
 ignore_duplicate_rows_when_importing = true  # If true, calculate the signature (hash) for each row when importing and avoid inserting the row if the signature already exists.
 slack_username = ""  # slack username for reporting errors when importing data.
 slack_channel = ""  # slack channel for reporting errors when importing data.
